More stdio formatted output <-> argument alignments in *printf().
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@4227 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/bogolexer.c
@@ -304,7 +304,7 @@ int main(int argc, char **argv)
 		fprintf(stdout, "%s\n", yylval->text);
 	    }
 	    else if (!quiet)
-		printf("get_token: %d \"%s\"\n", t, yylval->text);
+		printf("get_token: %d \"%s\"\n", (int)t, yylval->text);
 	}
     }
